updated the files from Sep. 2004 version to June 2005 version
Copy/paste of the files from June 2005 version to see the difference
between the Sep. 2004 version of Beat Luethi to the June 2005 version
(probably with some new interpolations, etc.). Unfortunately, the
documentation is not available, some reverse engineering is needed. Next
versions of this software were written in Visual C++ without graphics.